feat: Adicionando dinamismo na tela de perfil

+++ src/main/java/br/com/cadeaagua/api/repository/UsuarioRepository.java
@@ -8,8 +8,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface UsuarioRepository extends JpaRepository<Usuario, Long> {
+public interface UsuarioRepository extends JpaRepository<Usuario, Integer> {
     List<Usuario> findByEnderecoBairro(String bairro);
-    // Método para buscar por e-mail
     Optional<Usuario> findByEmail(String email);
 }
